my_sample_project/requirements.txt
django-crispy-forms
crispy-bootstrap5
django_extensions
python-decouple
Pillow